various refactorings
