+++ geolud-server-node/src/app.module.ts
@@ -4,11 +4,13 @@ import { ProjectController } from './projects/project.controller';
 import { ContactController } from './contact/contact.controller';
 import { MailService } from './mail/mail.service';
 import { Project } from './projects/project.entity';
-import { DataInitializerService } from './core/data-initializer/data-initializer.service';
-import { CONFIG } from './core/config/config';
+import { DataInitializerService } from './core/data-initializer.service';
+import { ConfigurationService } from './core/configuration.service';
+
+const appConfig = new ConfigurationService().init();
 
 @Module({
-  imports: [TypeOrmModule.forRoot(CONFIG.db), TypeOrmModule.forFeature([Project])],
+  imports: [TypeOrmModule.forRoot(appConfig.dbConfig), TypeOrmModule.forFeature([Project])],
   controllers: [ProjectController, ContactController],
   providers: [MailService, DataInitializerService],
 })
