--- geolud-server-node/src/core/config/config.types.ts
@@ -1,15 +1,0 @@
-import {MongoConnectionOptions} from 'typeorm/driver/mongodb/MongoConnectionOptions';
-
-
-export interface MailOptions {
-    host: string;
-    port: string;
-    username: string;
-    password: string;
-    to: string;
-}
-
-export interface Config {
-    db: MongoConnectionOptions;
-    mail: MailOptions;
-}
