+++ geolud-server-node/src/core/projects.ts
@@ -1,4 +1,4 @@
-import { Project } from '../../projects/project.entity';
+import { Project } from '../projects/project.entity';
 
 export const ALL_PROJECTS: Project[] = [
   {
